Fix: use correct bit mask for fixing argb alphas
The bit mask used in the conditional was 0xFC000000, which leaves two
bits unchecked. So, alpha values at 0x04 and below would be made solid.
The new mask only fixes the alpha value if it's truly zero.

Also made the literals hex so it's easier to see what's going on.