Add entity API
Instead of getting a table with a small amount of data when calling EnvironmentDetectorPeripheral.scanEntities(), UUIDs for entities are now the only thing that is returned. These can be used with the new entity API. For example, entity.getNBT(uuid) returns the NBT data of the entity with the given UUID.

The entity API itself can be disabled in config, as well as the individual functions within it that could be exploitative on servers.

A PlayerDetectorPeripheral.getPlayerUUID(username) function has also been added so that you can make use of the entity API when using a PlayerDetectorPeripheral.