[Backport to 1.19.2] add utf8 message support to chatboxes
adds optional "useUtf8" parameter, informing us that the program is utf8 aware, providing its strings as utf8,
This allows sending messages in the extended minecraft character set.

see the utf8 lib in cc, as well as the \u{xxxx} escape sequence.

also sends the events with a new utf8 message parameter.

This is a backport from #757